+++ clx/io/writer/file_writer.py
@@ -1,6 +1,7 @@
 from abc import ABC, abstractmethod
 from clx.io.writer.writer import Writer
 
+
 class FileWriter(Writer):
 
     @property