+++ clx/tests/test_factory.py
@@ -1,6 +1,5 @@
 import pytest
 import cudf
-import pandas
 import os
 
 from clx.io.factory.factory import Factory
@@ -36,15 +35,13 @@ fs_writer_config = {
     "output_format": "text",
 }
 
-# Temporarily changing over cuDF to pandasDF because of issue with equality checks.
-# Issue: https://github.com/rapidsai/cudf/issues/1750
 expected_df = cudf.DataFrame(
-    [
-        ("firstname", ["Emma", "Ava", "Sophia"]),
-        ("lastname", ["Olivia", "Isabella", "Charlotte"]),
-        ("gender", ["F", "F", "F"]),
-    ]
-).to_pandas()
+    {
+        "firstname": ["Emma", "Ava", "Sophia"],
+        "lastname": ["Olivia", "Isabella", "Charlotte"],
+        "gender": ["F", "F", "F"],
+    }
+)
 
 
 @pytest.mark.parametrize("kafka_config", [kafka_config])
@@ -92,9 +89,7 @@ def test_get_reader_text(test_input_base_path, expected_df):
     reader_from_factory = Factory.get_reader("fs", config)
     fetched_df = reader_from_factory.fetch_data()
 
-    # Temporarily changing over cuDF to pandasDF because of issue with equality checks.
-    # Issue: https://github.com/rapidsai/cudf/issues/1750
-    assert fetched_df.to_pandas().equals(expected_df)
+    assert fetched_df.equals(expected_df)
 
 
 @pytest.mark.parametrize("test_input_base_path", [test_input_base_path])
@@ -110,9 +105,7 @@ def test_get_reader_parquet(test_input_base_path, expected_df):
     reader_from_factory = Factory.get_reader("fs", config)
     fetched_df = reader_from_factory.fetch_data()
 
-    # Temporarily changing over cuDF to pandasDF because of issue with equality checks.
-    # Issue: https://github.com/rapidsai/cudf/issues/1750
-    assert fetched_df.to_pandas().equals(expected_df)
+    assert fetched_df.equals(expected_df)
 
 
 @pytest.mark.parametrize("test_input_base_path", [test_input_base_path])
@@ -128,6 +121,4 @@ def test_get_reader_orc(test_input_base_path, expected_df):
     reader_from_factory = Factory.get_reader("fs", config)
     fetched_df = reader_from_factory.fetch_data()
 
-    # Temporarily changing over cuDF to pandasDF because of issue with equality checks.
-    # Issue: https://github.com/rapidsai/cudf/issues/1750
-    assert fetched_df.to_pandas().equals(expected_df)
+    assert fetched_df.equals(expected_df)