+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## New Features
 
 ## Improvements
+ - PR #47 Update splunk workflow output
  - PR #43 Functionality to parse selected windows events
 
 ## Bug Fixes
 
+ - PR #45 More Kafka IO fixes
  - PR #44 Fix Kafka IO
  - PR #42 Include DNS parser in module
  - PR #41 Fix unit test
@@ -16,7 +18,9 @@
 # clx 0.10.0 (Date TBD)
 
 ## New Features
-
+ - PR #35 Added readthedocs
+ - PR #37 Add pytorch dependency.
+ - PR #37 Add DGA detection feature.
  - PR #14 Integrate repo with gpuCI
 
 ## Improvements